[lexical-utils][lexical-rich-text][lexical-plain-text] Refactor: Move eventFiles to @lexical/utils
Move eventFiles from @lexical/rich-text to @lexical/utils so both
rich-text and plain-text can share it. Rich-text re-exports it from
@lexical/utils for backwards compatibility. Plain-text now uses
eventFiles instead of an inlined file-transfer check.
