+++ packages/lexical-rich-text/src/index.ts
@@ -20,7 +20,6 @@ import type {
   NodeKey,
   NodeSelection,
   ParagraphNode,
-  PasteCommandType,
   RangeSelection,
   SerializedElementNode,
   Spread,
@@ -47,6 +46,7 @@ import {
   $getNearestBlockElementAncestorOrThrow,
   $handleIndentAndOutdent,
   addClassNamesToElement,
+  eventFiles,
   isHTMLElement,
   mergeRegister,
   objectKlassEquals,
@@ -513,29 +513,7 @@ async function onCutForRichText(
   );
 }
 
-// Clipboard may contain files that we aren't allowed to read. While the event is arguably useless,
-// in certain occasions, we want to know whether it was a file transfer, as opposed to text. We
-// control this with the first boolean flag.
-export function eventFiles(
-  event: DragEvent | PasteCommandType,
-): [boolean, File[], boolean] {
-  let dataTransfer: null | DataTransfer = null;
-  if (objectKlassEquals(event, DragEvent)) {
-    dataTransfer = event.dataTransfer;
-  } else if (objectKlassEquals(event, ClipboardEvent)) {
-    dataTransfer = event.clipboardData;
-  }
-
-  if (dataTransfer === null) {
-    return [false, [], false];
-  }
-
-  const types = dataTransfer.types;
-  const hasFiles = types.includes('Files');
-  const hasContent =
-    types.includes('text/html') || types.includes('text/plain');
-  return [hasFiles, Array.from(dataTransfer.files), hasContent];
-}
+export {eventFiles} from '@lexical/utils';
 
 function $isTargetWithinDecorator(target: HTMLElement): boolean {
   const node = $getNearestNodeFromDOMNode(target);
